Refatoradas entidades 'Category' e 'Product'

+++ product-backend/src/main/java/com/abutua/product_backend/services/ProductService.java
@@ -20,7 +20,7 @@ public class ProductService {
     @Autowired
     private CategoryService categoryService;
 
-    public Product getById(int id) {
+    public Product getById(long id) {
         Product product = productRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found"));
 
@@ -37,13 +37,13 @@ public class ProductService {
         return product;
     }
 
-    public void deleteById(int id) {
+    public void deleteById(long id) {
         Product product = getById(id);
         productRepository.delete(product);
 
     }
 
-    public void update(int id, Product productUpdate) {
+    public void update(long id, Product productUpdate) {
      
         Product product = getById(id);
